ajout de quelques fonctionnalite

+++ metier/src/main/java/com/groupeisi/dto/ClasseDto.java
@@ -11,7 +11,7 @@ import java.io.Serializable;
 public class ClasseDto implements Serializable {
 
     @XmlElement(name = "idClasse", required = true)
-    private Long id;
+    private Integer id; // Changement de Long à Integer
 
     @XmlElement(name = "className", required = true)
     private String className;
@@ -19,26 +19,24 @@ public class ClasseDto implements Serializable {
     @XmlElement(name = "description", required = true)
     private String description;
 
-    // Optionnel : On peut aussi référencer le secteur par son id (Long sectorId) si besoin,
-    // mais ici on reste simple et on ne met pas la relation entière.
-
     public ClasseDto() {
         super();
     }
 
-    public ClasseDto(Long id, String className, String description) {
+    // Le constructeur a été mis à jour pour utiliser Integer pour l'id
+    public ClasseDto(Integer id, String className, String description) {
         super();
         this.id = id;
         this.className = className;
         this.description = description;
     }
 
-    // Getters / Setters
-    public Long getId() {
+    // Getters / Setters mis à jour pour utiliser Integer pour l'id
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
